NOJIRA - allow easier testing of feature switches
All feature switches should behave the same. In order to be able to test the correct behaviour in APIs where there are multiple feature switches and reduce unnecessary boilerplate, this PR proposes a  `FeatureSwitchesBehaviour` trait that has test cases to check this in a general case (for 'enabled', 'released-in-production' and other flavours).

The FeatureSwitchesSpec uses this (to test the general purpose `isEnabled`/`isEnabledInProduction`) methods; and the API-specific BsasFeatureSwitchesSpec
does the same for those methods particular to BSAS.